\name{gcrma.exprSet}
\docType{data}
\alias{gcrma.exprSet}
\title{ExprSet resulting from gcrma processing of 3 HG-U133A chips}
\description{
	An exprSet object resulting from gcrma processing of three HG-U133A chips
	from the 28-chip affyCOMP II dataset, derived from the AffyMetrix 
	Latin Squares dataset. The gcrma.exprSet can be used as example input 
	for panp's pa.calls() function, and it is used for this purpose in
	the panp vignette.
}
\usage{data(gcrma.exprSet)}
\format{An exprSet object comprising three gcrma-processed HG-U133A chips}
\source{The 3 HG-U133A chips are the first three of affyCOMP II Latin Squares set}
\keyword{datasets}