\name{NSMPnames.hgu133plus2}
\docType{data}
\alias{NSMPnames.hgu133plus2}
\title{Negative Strand Matching Probeset (NSMP) names for HG-U133 Plus 2.0 chip type}
\description{
	This list of NSMP probeset names is used by pa.calls() 
	in calculating the NSMP expression distribution. It is
	specific to the chip type; the correct version is 
	automatically loaded for the detected chip type, if
	supported. The user does not load or use this
	data directly. 
}
\usage{data(NSMPnames.hgu133plus2)}
\format{A vector containing probeset names}
\source{October 2004 AffyMetrix chip annotation files, filtered, with outliers removed}
\keyword{datasets}