\name{test133a}
\docType{data}
\alias{test133a}
\title{Test HG-U133A exprSet for pa.calls}
\description{
	An example exprSet object from a small set of 
	HG-U133A chips, preprocessed with RMA. The pa.calls() 
	function can be executed on this exprSet, to exercise the
	function and generate P/A calls and related P-values.
	See ?help(pa.calls) for details. This exprSet is the one used
	in the help(pa.calls) example. 
}
\usage{data(test133a)}
\format{A 3-chip HG-U133A exprSet object (see the affy package for details 
on this object type) containing 3 samples, 22300 genes}
\source{A subset of the Latin Squares samples used in the affyCOMP II package}
\keyword{datasets}